# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Order API",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost/api",
            "description": "Local API Server"
        }
    ],
    "paths": {
        "/orders": {
            "get": {
                "tags": [
                    "Orders"
                ],
                "summary": "Список всех заказов",
                "description": "Получение списка заказов.",
                "operationId": "325fd3d55bbe9bb39645269928fa4983",
                "responses": {
                    "200": {
                        "description": "Успешный ответ",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Order"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Orders"
                ],
                "summary": "Создать заказ",
                "description": "Создание нового заказа.",
                "operationId": "445bab4f7dc9ea6fcd54f286c0209478",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Order"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Заказ создан",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Order"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/orders/{id}": {
            "get": {
                "tags": [
                    "Orders"
                ],
                "summary": "Получить заказ по ID",
                "description": "Получение конкретного заказа.",
                "operationId": "f088d34a2737782215f435f854e31de5",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID заказа",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Информация о заказе",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Order"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Заказ не найден"
                    }
                }
            },
            "put": {
                "tags": [
                    "Orders"
                ],
                "summary": "Обновить заказ",
                "description": "Обновление заказа.",
                "operationId": "54c2858108887d7a1252bcc4325158c2",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID заказа",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Order"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Заказ обновлен",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Order"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Orders"
                ],
                "summary": "Удалить заказ",
                "description": "Удаление заказа.",
                "operationId": "deb80b1c70721ac0812c9574580984c1",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID заказа",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Заказ удален"
                    }
                }
            }
        },
        "/products": {
            "get": {
                "tags": [
                    "Products"
                ],
                "summary": "Список товаров",
                "description": "Показать список всех товаров.",
                "operationId": "c1ff6f862214e1896c59cfe0491ce0e8",
                "responses": {
                    "200": {
                        "description": "Список товаров",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Product"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Products"
                ],
                "summary": "Создать новый товар",
                "description": "Создать новый товар.",
                "operationId": "399770d33fa4672f1216707d0e236763",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Product"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Товар создан",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Product"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/products/{id}": {
            "get": {
                "tags": [
                    "Products"
                ],
                "summary": "Получить конкретный товар",
                "description": "Показать конкретный товар.",
                "operationId": "b46f4aa73232d0368eee698a082342a4",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID товара",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Детали товара",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Product"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Товар не найден"
                    }
                }
            },
            "put": {
                "tags": [
                    "Products"
                ],
                "summary": "Обновить товар",
                "description": "Обновить информацию о товаре.",
                "operationId": "1419d479f9340e30e67fe21a2ca525b7",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID товара",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Product"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Товар обновлен",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Product"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Товар не найден"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Products"
                ],
                "summary": "Удалить товар",
                "description": "Удалить товар.",
                "operationId": "fe34b364a3f192305598b65d04d06b5d",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID товара",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Товар удален"
                    },
                    "404": {
                        "description": "Товар не найден"
                    }
                }
            }
        },
        "/profiles": {
            "get": {
                "tags": [
                    "Profiles"
                ],
                "summary": "Получить список профилей",
                "operationId": "313661758971eb804cf92949669fbc2e",
                "responses": {
                    "200": {
                        "description": "Список профилей",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Profile"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Profiles"
                ],
                "summary": "Создать новый профиль",
                "operationId": "22ac13592fe7c5d287d6186f134069fb",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Profile"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Профиль успешно создан",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Profile"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/profiles/{id}": {
            "get": {
                "tags": [
                    "Profiles"
                ],
                "summary": "Получить конкретный профиль по ID",
                "operationId": "6cef3d4208b8675f72b727102e18b631",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID профиля для получения",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Детали профиля",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Profile"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Профиль не найден"
                    }
                }
            },
            "put": {
                "tags": [
                    "Profiles"
                ],
                "summary": "Обновить конкретный профиль по ID",
                "operationId": "f78406d6d30c32faa5ec7cf84268b8e3",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID профиля для обновления",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Profile"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Профиль успешно обновлен",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Profile"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Профиль не найден"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Profiles"
                ],
                "summary": "Удалить конкретный профиль по ID",
                "operationId": "f9e09934fa7d892c46b82ff64403571a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID профиля для удаления",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Профиль успешно удален"
                    },
                    "404": {
                        "description": "Профиль не найден"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "Order": {
                "title": "Order",
                "required": [
                    "id",
                    "user_id",
                    "total_price",
                    "status"
                ],
                "properties": {
                    "id": {
                        "description": "ID заказа",
                        "type": "integer"
                    },
                    "user_id": {
                        "description": "ID пользователя",
                        "type": "integer"
                    },
                    "total_price": {
                        "description": "Общая стоимость заказа",
                        "type": "number",
                        "format": "float"
                    },
                    "status": {
                        "description": "Статус заказа",
                        "type": "string",
                        "example": "pending"
                    },
                    "products": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Product"
                        }
                    },
                    "user": {
                        "$ref": "#/components/schemas/User"
                    }
                },
                "type": "object"
            },
            "Product": {
                "title": "Product",
                "required": [
                    "id",
                    "name",
                    "price",
                    "stock"
                ],
                "properties": {
                    "id": {
                        "description": "ID товара",
                        "type": "integer"
                    },
                    "name": {
                        "description": "Название товара",
                        "type": "string"
                    },
                    "price": {
                        "description": "Цена товара",
                        "type": "number",
                        "format": "float"
                    },
                    "stock": {
                        "description": "Количество товара на складе",
                        "type": "integer"
                    },
                    "orders": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Order"
                        }
                    }
                },
                "type": "object"
            },
            "Profile": {
                "title": "Profile",
                "required": [
                    "id",
                    "user_id",
                    "address",
                    "phone"
                ],
                "properties": {
                    "id": {
                        "description": "ID профиля",
                        "type": "integer"
                    },
                    "user_id": {
                        "description": "ID пользователя, с которым связан профиль",
                        "type": "integer"
                    },
                    "address": {
                        "description": "Адрес пользователя",
                        "type": "string"
                    },
                    "phone": {
                        "description": "Телефон пользователя",
                        "type": "string"
                    },
                    "user": {
                        "$ref": "#/components/schemas/User"
                    }
                },
                "type": "object"
            },
            "User": {
                "title": "User",
                "required": [
                    "id",
                    "name",
                    "email"
                ],
                "properties": {
                    "id": {
                        "description": "ID пользователя",
                        "type": "integer"
                    },
                    "name": {
                        "description": "Имя пользователя",
                        "type": "string"
                    },
                    "email": {
                        "description": "Email пользователя",
                        "type": "string"
                    },
                    "orders": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Order"
                        }
                    },
                    "profile": {
                        "$ref": "#/components/schemas/Profile"
                    }
                },
                "type": "object"
            }
        }
    },
    "tags": [
        {
            "name": "Profiles",
            "description": "Операции, связанные с профилями пользователей"
        },
        {
            "name": "Orders",
            "description": "Orders"
        },
        {
            "name": "Products",
            "description": "Products"
        }
    ]
}